Struct stack::Stack

pub struct Stack { /* private fields */ }

A range of mapped memory designated for use as a task’s stack.

There is an unmapped guard page beneath the stack, which is a standard approach to detect stack overflow.

A stack is backed by and auto-derefs into MappedPages.

Implementations

impl Stack

pub fn top_unusable(&self) -> VirtualAddress

Returns the address just beyond the top of this stack, which is necessary for some hardware registers to use.

This address is not dereferenceable, the one right below it is. To get the highest usable address in this Stack, call top_usable()

pub fn top_usable(&self) -> VirtualAddress

Returns the highest usable address of this Stack, which is top_unusable() - sizeof(VirtualAddress)

pub fn bottom(&self) -> VirtualAddress

Returns the bottom of this stack, its lowest usable address.

pub fn from_pages(
guard_page: AllocatedPages,
stack_pages: MappedPages
) -> Result<Stack, (AllocatedPages, MappedPages)>

Creates a stack from its constituent parts: a guard page and a series of mapped pages.

Conditions

The guard_page must be at least one page (which is unmapped) and must contiguously precede the stack_pages. In other words, the beginning of stack_pages must come right after the end of guard_page.

The stack_pages must be mapped as writable.

If the conditions are not met, an Err containing the given guard_page and stack_pages is returned.

pub fn guard_page(&self) -> &PageRange

Returns the guard page(s) for this stack.

Guard pages are virtual pages that are reserved/owned by this stack but are not mapped, causing any access to them to result in a page fault.